requirement.txt
streamlit==1.40.1
google-generativeai==0.8.4
